+++ cmd/commands/get.go
@@ -18,14 +18,15 @@ var GetCmd = &cobra.Command{
 	Args:    cobra.MinimumNArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		// nolint:staticcheck
-		if err := database.DB.Update(
+		if err := database.DB.View( // ✅ Mudar Update para View
 			func(tx *nutsdb.Tx) error {
 				key := []byte(args[0])
 				content, err := tx.Get(database.Bucket, key)
-				must.Must(err, "GetCmd() - key not found or datababse must be empty.")
+				must.Must(err, "GetCmd() - key not found or database must be empty.")
 				fmt.Printf("%s\n", content)
 				return nil
 			}); err != nil {
+			must.Must(err, "GetCmd() - failed to retrieve key") // ✅ Tratar erro
 		}
 	},
 }
